// frontend/src/shared/api/generated/model/attemptResultDto.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * Quiz API
 * OpenAPI spec version: 1.0
 */
import type { TraitScoreDto } from './traitScoreDto';
import type { RecommendationDto } from './recommendationDto';

/**
 * Detailed result of a completed quiz attempt
 */
export interface AttemptResultDto {
  /** Calculated trait scores.
The key represents the trait code, the value represents the score.
 */
  traitScores?: TraitScoreDto[];
  /** List of recommended professions based on ML prediction */
  recommendations?: RecommendationDto[];
}